+++ src/backends/postgresql/vector-use-type.cpp
@@ -124,24 +124,24 @@ void postgresql_vector_use_type_backend::pre_use(indicator const * ind)
                 break;
             case x_long_long:
                 {
-                    std::vector<long long>* pv
-                        = static_cast<std::vector<long long>*>(data_);
-                    std::vector<long long>& v = *pv;
+                    std::vector<long>* pv
+                        = static_cast<std::vector<long>*>(data_);
+                    std::vector<long>& v = *pv;
 
                     std::size_t const bufSize
-                        = std::numeric_limits<long long>::digits10 + 3;
+                        = std::numeric_limits<long>::digits10 + 3;
                     buf = new char[bufSize];
                     snprintf(buf, bufSize, "%" LL_FMT_FLAGS "d", v[i]);
                 }
                 break;
             case x_unsigned_long_long:
                 {
-                    std::vector<unsigned long long>* pv
-                        = static_cast<std::vector<unsigned long long>*>(data_);
-                    std::vector<unsigned long long>& v = *pv;
+                    std::vector<unsigned long>* pv
+                        = static_cast<std::vector<unsigned long>*>(data_);
+                    std::vector<unsigned long>& v = *pv;
 
                     std::size_t const bufSize
-                        = std::numeric_limits<unsigned long long>::digits10 + 2;
+                        = std::numeric_limits<unsigned long>::digits10 + 2;
                     buf = new char[bufSize];
                     snprintf(buf, bufSize, "%" LL_FMT_FLAGS "u", v[i]);
                 }
